update var declaration

+++ src/python_routing_v02/fast_reach/fortran_wrappers.pxd
@@ -69,7 +69,7 @@ cdef extern from "pydiffusive.h":
                      double *ufhlt_f_g,
                      double *ufqlt_f_g,
                      int *frnw_col,
-                     int *frnw_g,
+                     double *frnw_g,
                      double *qlat_g,
                      double *ubcd_g,
                      double *dbcd_g,
